fix: Introduce classification threshold + 'other' keyword
Facing issues with non-invoice/license/bank statement docs (i.e.
false-positive rate), so this change uses 'other' as the keyword to
guide BERT to output the right class in these cases. In addition, we
currently take the max classification label when we should filter our
low-confidence results as 'other'/unknown to avoid FP rate. With these
changes, found that we are
no longer overfitting to positive classes by testing on misc PDFs, but at
the expense of lost accuracy (2% reduction). This can be further tuned.